Fix problem with match, when no arguments are provided
Match assumed that the minimum size of args will be always be > 2, being
args[0] the program name, argv[1] the command name, and arg[2:] the rest of the
options. The tests covers this, however they don't cover the case the user
don't provide any argument (in which case argv will contain only the program
name), which will derive into a out of index error.
I have just add a check to avoid this.

This problem should not occur for apply (which also manipulates this), since an
invalid range returns an empty list.